tests/wpt/tests/document-picture-in-picture/display-mode.https.html

<!DOCTYPE html>
<title>Test picture-in-picture display mode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>
<script>
promise_test(async (t) => {
  await test_driver.bless('request PiP window');
  const pipWindow = await documentPictureInPicture.requestWindow();
  await new Promise(requestAnimationFrame);
  assert_true(!!pipWindow.matchMedia('(display-mode: picture-in-picture)'.matches));
});
</script>
</body>
